integrate the avgtasktime calculation in the project to get executor duration of new job

+++ src/cn/hit/cst/ssl/bean/YARNHistoryJob.java
@@ -88,6 +88,6 @@ public abstract class YARNHistoryJob {
 		this.vcoreSec = vcoreSec;
 	}
 	
-	public abstract int getX();
+	public abstract Double getX();
 }
