+++ src/cn/hit/cst/ssl/utils/Sortable.java
@@ -1,5 +1,5 @@
 package cn.hit.cst.ssl.utils;
 
 public abstract class Sortable {
-	public abstract int getMyKey();
+	public abstract Double getMyKey();
 }
